# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: SkyHigh Core API
  version: "1.0.0"
  description: |
    REST API for SkyHigh Core digital check-in system.
    Covers authentication, flights, seat management, check-in, baggage, payments, and waitlist.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/flights/{flightId}/seats/{seatNumber}/waitlist:
    post:
      tags: [Waitlist]
      summary: Join waitlist for a specific seat
      operationId: joinWaitlist
      parameters:
        - name: flightId
          in: path
          required: true
          schema:
            type: string
        - name: seatNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaitlistJoinRequestDTO'
      responses:
        "201":
          description: Waitlist joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistResponseDTO'
  /api/v1/waitlist/{waitlistId}:
    delete:
      tags: [Waitlist]
      summary: Leave waitlist
      operationId: leaveWaitlist
      parameters:
        - name: waitlistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Left waitlist
  /api/v1/waitlist/{waitlistId}/position:
    get:
      tags: [Waitlist]
      summary: Get waitlist position
      operationId: getWaitlistPosition
      parameters:
        - name: waitlistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Waitlist position
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistPositionDTO'
  /api/v1/passengers/{passengerId}/waitlist:
    get:
      tags: [Waitlist]
      summary: Get all waitlist entries for passenger
      operationId: getPassengerWaitlist
      parameters:
        - name: passengerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Passenger waitlist entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WaitlistResponseDTO'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WaitlistJoinRequestDTO:
      type: object
      required: [passengerId]
      properties:
        passengerId:
          type: string
    WaitlistResponseDTO:
      type: object
      properties:
        waitlistId:
          type: integer
          format: int64
        passengerId:
          type: string
        flightId:
          type: string
        seatNumber:
          type: string
        position:
          type: integer
        status:
          $ref: '#/components/schemas/WaitlistStatus'
        joinedAt:
          type: string
          format: date-time
        notifiedAt:
          type: string
          format: date-time
          nullable: true
        assignedAt:
          type: string
          format: date-time
          nullable: true
        expiredAt:
          type: string
          format: date-time
          nullable: true
        message:
          type: string
    WaitlistPositionDTO:
      type: object
      properties:
        waitlistId:
          type: integer
          format: int64
        flightId:
          type: string
        seatNumber:
          type: string
        position:
          type: integer
        totalWaiting:
          type: integer
        status:
          $ref: '#/components/schemas/WaitlistStatus'
        message:
          type: string
...
    WaitlistStatus:
      type: string
      enum: [ACTIVE, NOTIFIED, ASSIGNED, CANCELLED, EXPIRED]
security:
  - bearerAuth: []